refactor: payment folders
Splitting Payment Repository/Service/Controller from Ticket to be more concise about what is doing what
