+++ src/controllers/index.ts
@@ -3,3 +3,4 @@ export * from './authentication-controller';
 export * from './events-controller';
 export * from './enrollments-controller';
 export * from './ticket-controller';
+export * from './payment-controller';
