+++ src/controllers/ticket-controller.ts
@@ -19,19 +19,3 @@ export async function findTicketByEnrollmentId(req: Request, res: Response) {
 
   res.send(ticket).status(httpStatus.OK);
 }
-
-export async function updatePayment(req: Request, res: Response) {
-  const { userId } = req.body;
-
-  await ticketService.updatePayment(userId);
-
-  res.sendStatus(httpStatus.OK);
-}
-
-export async function findPaymentByUserId(req: Request, res: Response) {
-  const { userId } = res.locals;
-
-  const payment = await ticketService.findPaymentByUserId(parseInt(userId));
-
-  res.send(payment).status(httpStatus.OK);
-}
