+++ src/services/index.ts
@@ -3,3 +3,4 @@ export * from './authentication-service';
 export * from './events-service';
 export * from './enrollments-service';
 export * from './ticket-service';
+export * from './payment-service';
