+++ src/services/payment-service/index.ts
@@ -1,0 +1,16 @@
+import paymentRepository from '@/repositories/payment-repository';
+
+export async function findPaymentByUserId(userId: number) {
+  return await paymentRepository.findPaymentByUserId(userId);
+}
+
+export async function updatePayment(userId: number) {
+  await paymentRepository.updatePayment(userId);
+}
+
+const paymentService = {
+  updatePayment,
+  findPaymentByUserId,
+};
+
+export default paymentService;
